Add files via upload
This is a program I created so that studying English vocabulary and common expressions for my C1 exam was more exciting. It makes use of a csv file called 'voc' which contains words or phrases in your mother tongue and in English, each pair or words in one line, separated by a comma. You start having three 'lives', and you can choose if you wish to translate from English to Spanish of viceversa. Each time you make a mistake, you are given a second opportunity and a hint (first letter of the asked word or phrase). If you make a second mistake, you lose one life. If you guess 5 words on a roll, you are rewarded with an extra life. Needless to say, the goal of this game is to spell all the words from you csv file correctly. I must say: it works!